backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
python-dotenv==1.0.0
google-generativeai==0.5.2
python-multipart==0.0.12
